[PIB-58] Moving file and refactor to formik
